requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
transformers>=4.30.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0